chore: Refactors OutOfStock component
